Do not use https to fetch cacert.org certificates
cacert.org uses its own class 3 PKI for https://cacert.org, which is not trusted by default.
This commit resolves the issue that the certificate cannot be downloaded via https, as the cacert
PKI is not trusted yet.

To ensure that noone injects a malicious certificate, we're checking the SHA checksum of the
downloaded certificate